Added functionality to the search bar in the books page
Added api endpoint to search for 20 books based on a string search term. Added services and hook method to front end to use the new search for books endpoint, also added a autocomplete search bar, so when you search for a book you can click on the reccomendations and be directed to the page of the book.