Book reccomendations appear on the search page
Added two more methods to the repository for books, get 8 of the highest rated books and get 8 of the books with the most reviews. Added the service methods to access the new api end points for the new methods, used the methods in the useEffect hooks then displayed the books in the book page and home page. Also styled some headers to make the pages look good.